Allow specifying the root-dir for translations when initializating.
Before, it was hard-coded to 'translations/...', which is *almost*
perfect, except it was relative to where the flask app was, so it was
doing 'api/translations/...', which doesn't exist.  Now I can specify
the right location in my code.

Test plan:
Added this to v1.py:

@route("/api/v1/test", methods=["GET"])
@api.auth.decorators.open_access
@jsonify
def i18n_test():
    import i18n
    return i18n._("Coach")

Also set up the KA code to properly specify the translations directory.

Then ran a dev-appserver and visited
localhost:8080/api/v1/test?lang=boxes and verified I saw boxes.