Get rid of the pkg_resources import in __init__.py.
We don't need it, as far as I can tell, and it causes appengine no end
of trouble in production: I was getting all these infinite redirects
when trying to import this file.  The fact we're doing weird stuff
with imports from third_party doesn't help

Test plan:
Did a non-default deploy and managed to successfully view the homepage.